fix: improvements to the RocketItemRenderer

+++ src/main/java/dev/galacticraft/mod/client/render/entity/rocket/RocketEntityRenderer.java
@@ -49,7 +49,7 @@ public class RocketEntityRenderer extends EntityRenderer<RocketEntity> {
         super.render(entity, yaw, partialTick, matrices, vertexConsumers, light);
         matrices.pushPose();
         Minecraft client = Minecraft.getInstance();
-        double rotationOffset = 1.6D;
+        float rotationOffset = 1.6F;
         matrices.translate(0, rotationOffset, 0);
         if (entity.getLaunchStage() == LaunchStage.IGNITED) {
             matrices.translate((entity.level().random.nextDouble() - 0.5D) * 0.1D, 0, (entity.level().random.nextDouble() - 0.5D) * 0.1D);
@@ -74,7 +74,7 @@ public class RocketEntityRenderer extends EntityRenderer<RocketEntity> {
         }
 
 //        RenderSystem.setShaderTexture(0, getTextureLocation(entity));
-        matrices.translate(0.0D, 0.4375D, 0.0D);
+        matrices.translate(0.0F, 0.4375F, 0.0F);
 
         Holder<? extends RocketPart<?, ?>> part = entity.engine();
         if (part != null) {
@@ -83,7 +83,7 @@ public class RocketEntityRenderer extends EntityRenderer<RocketEntity> {
             matrices.popPose();
         }
 
-        matrices.translate(0.0D, 0.5, 0.0D);
+        matrices.translate(0.0F, 0.5F, 0.0F);
 
         part = entity.booster();
         if (part != null) {
@@ -99,7 +99,7 @@ public class RocketEntityRenderer extends EntityRenderer<RocketEntity> {
             matrices.popPose();
         }
 
-        matrices.translate(0.0D, 1.0D, 0.0D);
+        matrices.translate(0.0F, 1.0F, 0.0F);
 
         part = entity.body();
         if (part != null) {
@@ -108,7 +108,7 @@ public class RocketEntityRenderer extends EntityRenderer<RocketEntity> {
             matrices.popPose();
         }
 
-        matrices.translate(0.0D, 1.75D, 0.0D);
+        matrices.translate(0.0F, 1.75F, 0.0F);
 
         part = entity.cone();
         if (part != null) {
